fix(discussion, comment, report): 토론 출력 되는 내용, 신고 하기 내용, 댓글 달기 오류 수정
- 댓글 달기 오류 수정
답글 달기 시 totalOrder가 부모 댓글 바로 밑의 숫자로 변하지 않는 문제 수정
- 토론, 신고 출력 내용 수정
토론에서는 신고를 위해 그 토론의 id값도 같이 출력을 해준다.
토론, 댓글 신고를 하나의 신고로 처리 하였으나, 편의를 위해 토론, 댓글 두 가지의 컨트롤러를 활용하여 신고하기를 분류함.

+++ src/main/java/com/undefinedus/backend/service/DiscussionServiceImpl.java
@@ -194,6 +194,7 @@ public class DiscussionServiceImpl implements DiscussionService {
         Discussion savedDiscussion = discussionRepository.save(discussion);
 
         DiscussionDetailResponseDTO discussionDetailResponseDTO = DiscussionDetailResponseDTO.builder()
+            .discussionId(discussionId)
             .bookTitle(discussionBook.getTitle())
             .memberName(discussion.getMember().getNickname())
             .title(savedDiscussion.getTitle())
@@ -260,25 +261,28 @@ public class DiscussionServiceImpl implements DiscussionService {
         DiscussionParticipant savedParticipant = discussionParticipantRepository.findByDiscussionAndMember(
             discussion, member).orElse(null);
 
-        if (savedParticipant == null) {
-            DiscussionParticipant discussionParticipant = DiscussionParticipant.builder()
-                .discussion(discussion)
-                .member(member)
-                .isAgree(true)
-                .build();
+        if (discussion.getStatus() == DiscussionStatus.PROPOSED) {
 
-            discussionParticipantRepository.save(discussionParticipant);
-        } else {
-            if (!savedParticipant.isAgree()) {
-
-                discussionParticipantRepository.deleteById(savedParticipant.getId());
-
+            if (savedParticipant == null) {
                 DiscussionParticipant discussionParticipant = DiscussionParticipant.builder()
                     .discussion(discussion)
                     .member(member)
                     .isAgree(true)
                     .build();
+
                 discussionParticipantRepository.save(discussionParticipant);
+            } else {
+                if (!savedParticipant.isAgree()) {
+
+                    discussionParticipantRepository.deleteById(savedParticipant.getId());
+
+                    DiscussionParticipant discussionParticipant = DiscussionParticipant.builder()
+                        .discussion(discussion)
+                        .member(member)
+                        .isAgree(true)
+                        .build();
+                    discussionParticipantRepository.save(discussionParticipant);
+                }
             }
         }
     }
@@ -296,19 +300,8 @@ public class DiscussionServiceImpl implements DiscussionService {
         DiscussionParticipant savedParticipant = discussionParticipantRepository.findByDiscussionAndMember(
             discussion, member).orElse(null);
 
-        if (savedParticipant == null) {
-            DiscussionParticipant discussionParticipant = DiscussionParticipant.builder()
-                .discussion(discussion)
-                .member(member)
-                .isAgree(false)
-                .build();
-
-            discussionParticipantRepository.save(discussionParticipant);
-        } else {
-            if (savedParticipant.isAgree()) {
-
-                discussionParticipantRepository.deleteById(savedParticipant.getId());
-
+        if (discussion.getStatus() == DiscussionStatus.PROPOSED) {
+            if (savedParticipant == null) {
                 DiscussionParticipant discussionParticipant = DiscussionParticipant.builder()
                     .discussion(discussion)
                     .member(member)
@@ -316,6 +309,19 @@ public class DiscussionServiceImpl implements DiscussionService {
                     .build();
 
                 discussionParticipantRepository.save(discussionParticipant);
+            } else {
+                if (savedParticipant.isAgree()) {
+
+                    discussionParticipantRepository.deleteById(savedParticipant.getId());
+
+                    DiscussionParticipant discussionParticipant = DiscussionParticipant.builder()
+                        .discussion(discussion)
+                        .member(member)
+                        .isAgree(false)
+                        .build();
+
+                    discussionParticipantRepository.save(discussionParticipant);
+                }
             }
         }
     }
